app/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2

pandas==2.2.2
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.4.2
xgboost==2.1.1

mlflow==2.14.1
python-dotenv==1.0.1

boto3==1.34.162